% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_school_year}
\alias{get_school_year}
\title{Get School Year from Date}
\usage{
get_school_year(date)
}
\arguments{
\item{date}{A Date or POSIXct object}
}
\value{
A character string in the format "YYYY/YYYY+1". The school year is considered
to start in August, so dates before August are assigned to the previous year.
Will error if input is not a Date or POSIXct object.
}
\description{
Convert a date into a school year format (e.g., "2023/2024")
}
