% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{left_join_load}
\alias{left_join_load}
\title{Load and Left Join Prepared Data}
\usage{
left_join_load(
  data_main,
  data_to_join_name,
  ...,
  .path = NULL,
  .config_data_path = "data_prepared_dir"
)
}
\arguments{
\item{data_main}{A data frame to join data to.}

\item{data_to_join_name}{A string specifying the name of the RDS file (without extension).}

\item{...}{Arguments passed to dplyr::left_join.}

\item{.path}{Optional. A string specifying the path to load data from.}

\item{.config_data_path}{Optional. A string specifying the config key for the data path.}
}
\value{
A data frame containing the left join result of \code{data_main} with the loaded data.
Will error if the main data is NULL or if the file to join cannot be found.
}
\description{
Load a prepared dataset from disk and left join it with an existing dataset
}
