% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{Load Data File}
\usage{
load_data(
  config_key,
  ...,
  filename = NULL,
  path = NULL,
  config_data_path = NULL
)
}
\arguments{
\item{config_key}{Character string specifying the configuration key to lookup file information}

\item{...}{Additional arguments passed to readr::read_delim}

\item{filename}{Optional character string specifying the name of the file to read.
If NULL, filename is retrieved from config using config_key.}

\item{path}{Optional character string specifying the path to read from.
If NULL, path is retrieved from config using config_raw_data.}

\item{config_data_path}{Character string specifying the config key for raw data directory.
Defaults to NULL.}
}
\value{
A tibble containing the raw data from the specified file
}
\description{
Reads a raw data file based on configuration settings or direct file specifications.
Uses config package to retrieve file paths and names if not directly specified.
}
\details{
The function will first try to use provided filename and path.
If these are NULL, it will look up values in the config file using config_key
and config_raw_data respectively.
}
