% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{read_starterset_monthly_a05}
\alias{read_starterset_monthly_a05}
\title{Read A05 Monthly File}
\usage{
read_starterset_monthly_a05(filename)
}
\arguments{
\item{filename}{A single string specifying the file to read}
}
\value{
A tibble containing the processed A05 data with standardized column names
}
\description{
Read and process an A05 monthly data file from the education starterset
}
