% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_analysed}
\alias{save_analysed}
\title{Save Analyzed Data to RDS File}
\usage{
save_analysed(
  data,
  ...,
  filename = NULL,
  path = NULL,
  config_data_path = "data_analysed_dir"
)
}
\arguments{
\item{data}{An R object to save.}

\item{...}{Additional arguments passed to saveRDS().}

\item{filename}{Optional. A string for the output filename. If NULL, uses the input object name.}

\item{path}{Optional. A string specifying the output path.}

\item{config_data_path}{Optional. A string specifying the config key for the data directory.}
}
\value{
Invisibly returns the path where the file was saved. Creates directory if it
doesn't exist. Will error if the config package isn't available and no filename
is provided.
}
\description{
Save data to an RDS file in the analyzed data directory
}
