% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_combined}
\alias{save_combined}
\title{Save Combined Data to RDS}
\usage{
save_combined(
  data,
  ...,
  filename = NULL,
  path = NULL,
  config_data_path = "data_combined_dir"
)
}
\arguments{
\item{data}{The data object to save.}

\item{...}{Additional arguments passed to saveRDS().}

\item{filename}{Optional. A string for the output filename. If NULL, uses the data object's name.}

\item{path}{Optional. A string specifying the save location.}

\item{config_data_path}{Optional. A string specifying the config key for the save location.}
}
\value{
Invisibly returns the full path to the saved file. Creates directory if it
doesn't exist.
}
\description{
Save a data object to an RDS file in a specified directory
}
