% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_transformed_and_comment}
\alias{save_transformed_and_comment}
\title{Save Transformed Data with Optional Comments}
\usage{
save_transformed_and_comment(data, ..., filename = NULL, add_comment = TRUE)
}
\arguments{
\item{data}{A data frame or tibble to save.}

\item{...}{Currently unused; must be empty.}

\item{filename}{Optional. A string specifying the filename. If NULL, uses the data argument name.}

\item{add_comment}{Optional. A logical indicating whether to add the filename to comments.}
}
\value{
The data, invisibly, after saving it to disk.
}
\description{
Save a transformed dataset while optionally adding its filename to its comments
}
